+++ iti-react/Components/AddressDisplay.tsx
@@ -1,4 +1,8 @@
 ﻿import React from 'react'
+import {
+    formatAddressLine3,
+    isCanadianPostalCode
+} from '@interface-technologies/iti-react-core'
 
 export interface AddressDisplayAddress {
     line1: string
@@ -6,42 +10,6 @@ export interface AddressDisplayAddress {
     city: string
     state: string
     postalCode: string
-}
-
-function normalizePostalCode(postalCode: string) {
-    return postalCode.replace(' ', '').replace('-', '')
-}
-
-function isCanadian(address: AddressDisplayAddress) {
-    return normalizePostalCode(address.postalCode).length === 6
-}
-
-export function formatPostalCode(postalCode: string) {
-    postalCode = normalizePostalCode(postalCode)
-
-    switch (postalCode.length) {
-        case 9:
-            return postalCode.substr(0, 5) + '-' + postalCode.substr(5)
-        case 6:
-            // Canadian postal code
-            return postalCode.substr(0, 3) + ' ' + postalCode.substr(3)
-        default:
-        case 5:
-            return postalCode
-    }
-}
-
-function getLine3(address: AddressDisplayAddress) {
-    const postalCode = formatPostalCode(address.postalCode)
-
-    // building the line3 string this way because all fields are nullable
-    const stateZipParts = [address.state, postalCode].filter(s => !!s)
-    const stateZip = stateZipParts.join(' ')
-
-    let line3Parts = [address.city, stateZip].filter(s => !!s)
-
-    const separator = !isCanadian(address) ? ', ' : ' '
-    return line3Parts.join(separator)
 }
 
 interface AddressDisplayProps {
@@ -57,8 +25,8 @@ export function AddressDisplay(props: AddressDisplayProps) {
         <div className="address-display">
             {address.line1 && <p>{address.line1}</p>}
             {address.line2 && <p>{address.line2}</p>}
-            <p>{getLine3(address)}</p>
-            {isCanadian(address) && <p>Canada</p>}
+            <p>{formatAddressLine3(address)}</p>
+            {isCanadianPostalCode(address.postalCode) && <p>Canada</p>}
         </div>
     )
 }